Fixed wrong string length being passed
Fixed two occassions where I was not passing the correct string length
to the print function. One was off by 1 and the other was passing in the
allocated size not the string length

Changed end of string check condition as we could early leave the loop
early if two null characters happened right after each other. Moved do
while into a for loop.